Move tests to correct module
These tests aren't for the nav.metrics.threshold module, they're for
the nav.metrics.graphs module, so this moves them to the correct test
module.
